+++ sdk/dotnet/Pulumi/Deployment/GrpcMonitor.cs
@@ -1,6 +1,7 @@
 // Copyright 2016-2020, Pulumi Corporation
 
 using System.Threading.Tasks;
+using System.Collections.Generic;
 using Grpc.Core;
 using Pulumirpc;
 
@@ -12,7 +13,10 @@ namespace Pulumi
 
         public GrpcMonitor(string monitor)
         {
-            this._client = new ResourceMonitor.ResourceMonitorClient(new Channel(monitor, ChannelCredentials.Insecure));
+            // maxRpcMessageSize raises the gRPC Max Message size from `4194304` (4mb) to `419430400` (400mb)
+            var maxRpcMessageSize = 400 * 1024 * 1024;
+            var grpcChannelOptions = new List<ChannelOption> { new ChannelOption(ChannelOptions.MaxReceiveMessageLength, maxRpcMessageSize)};
+            this._client = new ResourceMonitor.ResourceMonitorClient(new Channel(monitor, ChannelCredentials.Insecure, grpcChannelOptions));
         }
         
         public async Task<InvokeResponse> InvokeAsync(InvokeRequest request)