+++ sdk/go/common/resource/plugin/plugin.go
@@ -59,9 +59,6 @@ type plugin struct {
 // pluginRPCConnectionTimeout dictates how long we wait for the plugin's RPC to become available.
 var pluginRPCConnectionTimeout = time.Second * 10
 
-// pluginRPCMaxMessageSize raises the gRPC Max Message size from `4194304` to `419430400`
-var pluginRPCMaxMessageSize = 1024 * 1024 * 400
-
 // A unique ID provided to the output stream of each plugin.  This allows the output of the plugin
 // to be streamed to the display, while still allowing that output to be sent a small piece at a
 // time.
@@ -180,13 +177,13 @@ func newPlugin(ctx *Context, pwd, bin, prefix string, args, env []string) (*plug
 	plug.stdoutDone = stdoutDone
 	go runtrace(plug.Stdout, false, stdoutDone)
 
-	// We want to increase the default message size as per pulumi/pulumi#2319
-	messageSizeOpts := grpc.WithDefaultCallOptions(grpc.MaxCallRecvMsgSize(pluginRPCMaxMessageSize))
-
 	// Now that we have the port, go ahead and create a gRPC client connection to it.
-	conn, err := grpc.Dial("127.0.0.1:"+port, grpc.WithInsecure(), grpc.WithUnaryInterceptor(
-		rpcutil.OpenTracingClientInterceptor(),
-	), messageSizeOpts)
+	conn, err := grpc.Dial(
+		"127.0.0.1:"+port,
+		grpc.WithInsecure(),
+		grpc.WithUnaryInterceptor(rpcutil.OpenTracingClientInterceptor()),
+		rpcutil.GrpcChannelOptions(),
+	)
 	if err != nil {
 		return nil, errors.Wrapf(err, "could not dial plugin [%v] over RPC", bin)
 	}
@@ -244,7 +241,7 @@ func execPlugin(bin string, pluginArgs []string, pwd string, env []string) (*plu
 	// Flow the logging information if set.
 	if logging.LogFlow {
 		if logging.LogToStderr {
-			args = append(args, "-logtostderr")
+			args = append(args, "--logtostderr")
 		}
 		if logging.Verbose > 0 {
 			args = append(args, "-v="+strconv.Itoa(logging.Verbose))